fix: Music folders containing OPUS format files cannot be recognized
log: The logic for dragging and dropping folders is fundamentally sound. It uses QDirIterator to recursively scan directories and filters supported file formats based on m_supportedSuffixs. The issue lies in the fact that the OPUS format was not previously included in the supported list. Consequently, folders containing OPUS files are ignored during the scanning process.

bug: https://pms.uniontech.com/bug-view-332161.html